pios_i2c: combine F3 and F4 implementations
So far in a naive way-- throwing all the contents into one file, dumb
combinations/transform of state machine.  Further transformations will
hopefully make this clean.

Tested OK on Sparky2 baro + sparky1 sensors.
