+++ flight/PiOS/STM32/inc/pios_i2c_priv.h
@@ -75,6 +75,14 @@ enum i2c_adapter_state {
 	I2C_STATE_W_LAST_TXN_ADDR,
 	I2C_STATE_W_LAST_TXN_MIDDLE,
 	I2C_STATE_W_LAST_TXN_LAST,
+
+	// F4
+	I2C_STATE_W_MORE_TXN_PRE_LAST,
+	I2C_STATE_W_MORE_TXN_POST_LAST,
+	I2C_STATE_W_MORE_TXN_PRE_MIDDLE,
+	I2C_STATE_W_LAST_TXN_PRE_MIDDLE,
+	I2C_STATE_W_LAST_TXN_PRE_LAST,
+	I2C_STATE_W_LAST_TXN_POST_LAST,
 	
 	I2C_STATE_NACK,
 	
@@ -82,15 +90,31 @@ enum i2c_adapter_state {
 };
 
 enum i2c_adapter_event {
+	I2C_EVENT_AUTO,
 	I2C_EVENT_START,
-	I2C_EVENT_RECEIVER_BUFFER_NOT_EMPTY,
-	I2C_EVENT_TRANSMIT_BUFFER_EMPTY,
-	I2C_EVENT_TRANSFER_COMPLETE,
 	I2C_EVENT_STOP,
 	I2C_EVENT_NACK,
 	I2C_EVENT_BUS_ERROR,
 	I2C_EVENT_STOPPED,
-	I2C_EVENT_AUTO,
+	
+	/* F3 Specific */
+	I2C_EVENT_RECEIVER_BUFFER_NOT_EMPTY,
+	I2C_EVENT_TRANSMIT_BUFFER_EMPTY,
+	I2C_EVENT_TRANSFER_COMPLETE,
+
+	/* F4 Specific */
+	I2C_EVENT_R_MORE_TXN_STARTED,
+	I2C_EVENT_W_MORE_TXN_STARTED,
+	I2C_EVENT_R_LAST_TXN_STARTED,
+	I2C_EVENT_W_LAST_TXN_STARTED,
+	I2C_EVENT_ADDR_SENT_LEN_EQ_0,
+	I2C_EVENT_ADDR_SENT_LEN_EQ_1,
+	I2C_EVENT_ADDR_SENT_LEN_EQ_2,
+	I2C_EVENT_ADDR_SENT_LEN_GT_2,
+	I2C_EVENT_TRANSFER_DONE_LEN_EQ_0,
+	I2C_EVENT_TRANSFER_DONE_LEN_EQ_1,
+	I2C_EVENT_TRANSFER_DONE_LEN_EQ_2,
+	I2C_EVENT_TRANSFER_DONE_LEN_GT_2,
 
 	I2C_EVENT_NUM_EVENTS	/* Must be last */
 };
